requirements.txt
Pillow
requests
Kivy
